10: Creates basic point damage framework
Four components are created:  Damageable, Destroyable, Health, and a Point
Damage type.

Damageable is reserved for things that can be hit by weapons/debris.

Destroyable is reserved for things that can despawn (player ship, lasers,
enemies, etc.).

Health is reserved solely for tracking health; it does not know/care about the
type of damage or what should happen to the object when its health reaches
zero.

Point Damage is used for single target damage.  A laser is a Point Damage since
the laser hits the object and then despawns itself.