Fix ship orientation
I made the same mistake with the ship as I did with the lasers.  I rotated the
whole game object so that y was forward, which causes problems.  This commit
adds a basic player mesh (just a triangle) and reorients the player game object
so that the forward vector works as intended.