Refactor handling of zero health
Now the Health Stat emits an event when the health changes that can be
subscribed to.  This way, we create a behavior that is "kill this thing when
its health is zero" rather than having multiple responsibilities for a single
attribute such as damageable.

Also fixes prefabs to work with this new system.