24: Adds object pool data component
An object pool holds a number of Game Objects that can be fetched to avoid
having to destroy and reinstantiate reusable objects over and over again.  E.G.
lasers.  Instead of destroying and spawning them, just get a laser prefab from
the laser object pool, use it, then deactivate it so the next client that wants
a laser can have it.