25: Creates basic single laser weapon
Weapons are a prefab combination of a Firable Attribute and a set of children
transforms for the locations of the projectile spawns.  The Firable Attribute
has a list of transforms (that are set in the prefab via the children
transforms), and it handles spawning entities from the associated object pool.

The object pools are linked via tags.  The weapon must be tagged with the same
tag used for the corresponding object pool.  E.G. both the Laser Object Pool
and any weapon that uses that pool must share the same tag.  For this commit,
the tag of "Laser" is used.

This commit also creates a basic single-fire laser weapon to test functionality.

The ship movement component is also changed to allow defining starting weapons
for the ship.  This is for setting the start weapons on init, but also for just
debugging other weapons in the future.