34: Ensure components are found on init if not provided
Sometimes a user will want to perform a specific action that is not handled by
default within the behavior component.  To fix this, all components that
behaviors/attributes depend on are now public.

However, to save people from having to manually hook up literally everything,
each component will search for things it depends on during init if not
provided.