33: Adds elapsed lifetime interface
If something depends on how long an object has been alive for, they can add a
component to it which will keep track of how long it has been enabled.  Since
we're dealing with scriptable behaviors, this also introduces a Seekable
interface.  If you want to rewind or fast-forward an object's behavior, you can
do that.