DOC: Add Game Screen UI mockup
A quick layout of how elements will be placed in the game screen is created.  It
is presented in 16:9 aspect ratio, so the relative sizes/positions of elements
on the screen are correct and reflect how the final product will look.  The
colors, font, etc. do not reflect the final product.  Rather, this is simply a
guide for how elements will be positioned on the screen.

This mockup also includes where the mobile controls will be.  On desktop,
obviously, these controls will not be present.